Rename files according to cpp styles.
